# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "AgentSummary": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "credentials": {
            "items": {
              "$ref": "#/components/schemas/CredentialLink"
            },
            "title": "Credentials",
            "type": "array"
          },
          "event_count": {
            "default": 0,
            "title": "Event Count",
            "type": "integer"
          },
          "framework": {
            "default": "",
            "title": "Framework",
            "type": "string"
          },
          "last_seen": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          },
          "session_count": {
            "default": 0,
            "title": "Session Count",
            "type": "integer"
          },
          "status": {
            "default": "active",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "agent_id"
        ],
        "title": "AgentSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SessionDetail": {
        "properties": {
          "blast_radius": {
            "items": {
              "$ref": "#/components/schemas/BlastResource"
            },
            "title": "Blast Radius",
            "type": "array"
          },
          "read_only": {
            "default": false,
            "title": "Read Only",
            "type": "boolean"
          },
          "session": {
            "$ref": "#/components/schemas/SessionSummary"
          },
          "timeline": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TimelineEvent"
                },
                {
                  "$ref": "#/components/schemas/GapMarker"
                }
              ]
            },
            "title": "Timeline",
            "type": "array"
          },
          "verification": {
            "$ref": "#/components/schemas/VerifyResult"
          }
        },
        "required": [
          "session",
          "timeline",
          "blast_radius",
          "verification"
        ],
        "title": "SessionDetail",
        "type": "object"
      },
      "SessionSummary": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "ended_at": {
            "title": "Ended At",
            "type": "string"
          },
          "error_count": {
            "title": "Error Count",
            "type": "integer"
          },
          "event_count": {
            "title": "Event Count",
            "type": "integer"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "started_at": {
            "title": "Started At",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "agent_id",
          "started_at",
          "ended_at",
          "event_count",
          "error_count"
        ],
        "title": "SessionSummary",
        "type": "object"
      },
...
      "ShareRequest": {
        "properties": {
          "expires_in_hours": {
            "default": 72,
            "maximum": 720.0,
            "minimum": 1.0,
            "title": "Expires In Hours",
            "type": "integer"
          }
        },
        "title": "ShareRequest",
        "type": "object"
      },
...
      "TimelineEvent": {
        "properties": {
          "credential": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CredentialLink"
              },
              {
                "type": "null"
              }
            ]
          },
          "credential_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credential Ref"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "kind": {
            "const": "event",
            "default": "event",
            "title": "Kind",
            "type": "string"
          },
          "operation": {
            "title": "Operation",
            "type": "string"
          },
          "outcome": {
            "title": "Outcome",
            "type": "string"
          },
          "payload": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          },
          "payload_truncated": {
            "title": "Payload Truncated",
            "type": "boolean"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "redactions": {
            "items": {
              "type": "string"
            },
            "title": "Redactions",
            "type": "array"
          },
          "resource_refs": {
            "items": {
              "type": "string"
            },
            "title": "Resource Refs",
            "type": "array"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "ts": {
            "title": "Ts",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "session_id",
          "seq",
          "ts",
          "source",
          "event_type",
          "operation",
          "provider",
          "target",
          "outcome",
          "duration_ms",
          "credential",
          "credential_ref",
          "resource_refs",
          "payload",
          "payload_truncated",
          "redactions"
        ],
        "title": "TimelineEvent",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Leaflyst API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/replay/agents": {
      "get": {
        "operationId": "agents_v1_replay_agents_get",
        "parameters": [
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-abx-admin-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Abx-Admin-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentSummary"
                  },
                  "title": "Response Agents V1 Replay Agents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Agents"
      }
    },
    "/v1/replay/agents/{agent_id}/sessions": {
      "get": {
        "operationId": "agent_sessions_v1_replay_agents__agent_id__sessions_get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-abx-admin-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Abx-Admin-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionSummary"
                  },
                  "title": "Response Agent Sessions V1 Replay Agents  Agent Id  Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Agent Sessions"
      }
    },
    "/v1/replay/credentials/{credential_id}/events": {
      "get": {
        "operationId": "credential_events_v1_replay_credentials__credential_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "credential_id",
            "required": true,
            "schema": {
              "title": "Credential Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-abx-admin-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Abx-Admin-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TimelineEvent"
                  },
                  "title": "Response Credential Events V1 Replay Credentials  Credential Id  Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Credential Events"
      }
    },
    "/v1/replay/sessions/{session_id}": {
      "get": {
        "operationId": "session_detail_v1_replay_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-abx-admin-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Abx-Admin-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Session Detail"
      }
    },
    "/v1/replay/sessions/{session_id}/blast-radius.csv": {
      "get": {
        "operationId": "blast_radius_csv_v1_replay_sessions__session_id__blast_radius_csv_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-abx-admin-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Abx-Admin-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Blast Radius Csv"
      }
    },
    "/v1/replay/sessions/{session_id}/blast-radius.json": {
      "get": {
        "operationId": "blast_radius_json_v1_replay_sessions__session_id__blast_radius_json_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-abx-admin-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Abx-Admin-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BlastResource"
                  },
                  "title": "Response Blast Radius Json V1 Replay Sessions  Session Id  Blast Radius Json Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Blast Radius Json"
      }
    },
    "/v1/replay/sessions/{session_id}/share": {
      "post": {
        "operationId": "create_share_v1_replay_sessions__session_id__share_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-abx-admin-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Abx-Admin-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Share"
      }
    },
    "/v1/replay/shared/{token}": {
      "get": {
        "operationId": "shared_session_v1_replay_shared__token__get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Shared Session"
      }
    },
...
  }
}